# util: set checkpoint during Sort for SortedRowContainer