--- a/util/chunk/row_container.go
+++ b/util/chunk/row_container.go
@@ -489,11 +489,16 @@ const SignalCheckpointForSort uint = 10240
 
 // keyColumnsLess is the less function for key columns.
 func (c *SortedRowContainer) keyColumnsLess(i, j int) bool {
-	if c.timesOfRowCompare == SignalCheckpointForSort {
+	if c.timesOfRowCompare >= SignalCheckpointForSort {
 		// Trigger Consume for checking the NeedKill signal
 		c.memTracker.Consume(1)
 		c.timesOfRowCompare = 0
 	}
+	failpoint.Inject("SignalCheckpointForSort", func(val failpoint.Value) {
+		if val.(bool) {
+			c.timesOfRowCompare += 1024
+		}
+	})
 	c.timesOfRowCompare++
 	rowI := c.m.records.inMemory.GetRow(c.ptrM.rowPtrs[i])
 	rowJ := c.m.records.inMemory.GetRow(c.ptrM.rowPtrs[j])